Revert create_launcher fix.
